backend/requirements.txt
fastapi==0.116.1
pydantic==2.11.7
sqlalchemy==2.0.42
uvicorn==0.35.0
mysql-connector-python==9.4.0
redis==5.0.1
celery==5.3.4
google-generativeai==0.3.2
weaviate-client==3.25.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jsonschema==4.21.1
python-dotenv==1.0.0
alembic==1.13.0